text-to-speach/requirements.txt
fastapi
uvicorn[standard]  
python-dotenv
cartesia